+++ src/shared/components/index.ts
@@ -44,4 +44,4 @@ export * from './LegalText'
 export * from './Loader'
 export * from './AnimatedError'
 export * from './EmptyFallback'
-export * from './GridLayout/GridLayout'
+export * from './LayoutGrid/LayoutGrid'
