+++ src/views/playground/Playgrounds/GridTesting.tsx
@@ -1,12 +1,12 @@
 import styled from '@emotion/styled'
 import React from 'react'
 
-import { GridItem, GridLayout } from '@/shared/components'
+import { GridItem, LayoutGrid } from '@/shared/components'
 
 export const GridTesting = () => {
   return (
     <div>
-      <GridLayout>
+      <LayoutGrid>
         <StyledGridItem></StyledGridItem>
         <StyledGridItem></StyledGridItem>
         <StyledGridItem></StyledGridItem>
@@ -16,7 +16,7 @@ export const GridTesting = () => {
           colStart={{ xlarge: 7 }}
           colSpan={{ small: 2, medium: 3, large: 4, xlarge: 5, xxlarge: 6 }}
         ></StyledGridItem>
-      </GridLayout>
+      </LayoutGrid>
     </div>
   )
 }
